# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Summit / Quadient / Sage 500 AP Integration API
  version: 1.1.0
  description: |
    API for Quadient Beanworks to retrieve Sage 500 data and send invoice information back to Summit.

    Current outbound GET endpoints:
    - Health / diagnostics
    - Companies
    - G/L Accounts
    - Customers
    - Vendors
    - Purchase Orders
    - Receipts / Receiving Info
    - Open AP Invoices
    - Customer Statements

    Current inbound POST endpoints:
    - Quadient invoice staging for Sage 500 Data Import Manager processing

    Notes:
    - Date/time values are returned from Sage 500.
    - Incremental endpoints support updatedSince where applicable.
    - Quadient is responsible for tracking the last successful sync date.
    - API key authentication uses the x-api-key request header.
    - The inbound invoice endpoint stores invoice data in staging tables only. It does not write directly to Sage 500 voucher tables.
servers:
  - url: http://localhost:3000
    description: Local testing endpoint
security:
  - ApiKeyAuth: []
...
paths:
...
  /db-test:
    get:
      summary: Test database connectivity
      description: Tests whether the API can connect to the Sage 500 SQL Server database.
      tags:
        - Diagnostics
      responses:
        '200':
          description: Database connection successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbTestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /companies:
    get:
      summary: Get companies
      description: Returns companies available from Sage 500.
      tags:
        - Reference
      responses:
        '200':
          description: Companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /gl-accounts:
    get:
      summary: Get G/L account list
      tags:
        - GL Accounts
      responses:
        '200':
          description: G/L accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GLAccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /customers:
    get:
      summary: Get customer list
      tags:
        - Customers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /vendors:
    get:
      summary: Get vendor list
      description: Returns vendors from Sage 500. VendKey should be used for Sage/DIM writeback matching when available.
      tags:
        - Vendors
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/UpdatedSince'
      responses:
        '200':
          description: Vendors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /purchase-orders:
    get:
      summary: Get purchase orders
      description: Returns purchase orders grouped by header, with line items nested under each purchase order.
      tags:
        - Purchase Orders
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Purchase orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /receipts:
    get:
      summary: Get receiving information
      description: Returns receipt / receiving information from Sage 500.
      tags:
        - Receipts
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Receipts / receiving info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /invoices:
    get:
      summary: Get open AP invoices
      description: Returns open AP invoices from Sage 500. Open is currently determined by non-zero balance.
      tags:
        - AP Invoices
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Open AP invoices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /statements:
    get:
      summary: Get customer statement-style open AR data
      description: Returns open customer invoice/balance data used to generate statement-style information.
      tags:
        - AR Statements
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - name: customerId
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Statement data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /quadient/invoice:
    post:
      summary: Receive invoice from Quadient Beanworks
      description: |
        Receives one 3-way matched invoice per request from Quadient Beanworks and stores it in staging tables for Sage 500 Data Import Manager processing.

        This endpoint does not write directly to Sage 500 voucher tables. After staging, Net@Work / Sage DIM can import from the DIM-facing views and then update the staging status.
      tags:
        - Quadient Inbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuadientInvoiceRequest'
            example:
              invoiceNumber: TEST-BEAN-1001
              vendorKey: 123
              vendorId: TESTVENDOR
              invoiceDate: '2026-05-07'
              dueDate: '2026-06-06'
              memo: Test invoice from Quadient API endpoint
              beanworksInvoiceUrl: https://example.com/beanworks/invoice/TEST-BEAN-1001
              currency: USD
              companyId: '001'
              totalAmount: 125.00
              lines:
                - lineNumber: 1
                  itemKey: 456
                  itemId: TESTITEM
                  unitCost: 12.50
                  quantity: 10
                  unitMeasure: EA
                  unitMeasKey: 1
                  lineAmount: 125.00
                  poKey: 10001
                  poNumber: PO-TEST-001
                  poLineKey: 20001
                  poLineNumber: 1
                  rcvrLineKey: 30001
                  description: Test invoice line
                  sTaxClassKey: null
                  department: TEST
                  costCenter: TEST
      responses:
        '201':
          description: Invoice received and staged for DIM import
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuadientInvoiceResponse'
              example:
                status: received
                processingStatus: ReadyForDIM
                stagingId: 101
                invoiceNumber: TEST-BEAN-1001
                vendorKey: 123
                vendorId: TESTVENDOR
                lineCount: 1
        '400':
          description: Payload validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                error: VALIDATION_FAILED
                message: Invoice payload failed validation
                details:
                  - invoiceNumber is required
                  - lines[0].rcvrLineKey is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Invoice staging failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INVOICE_RECEIVE_FAILED
                message: SQL insert failed
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    UpdatedSince:
      name: updatedSince
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Return records updated on or after this date/time where supported.
      example: '2026-05-01T00:00:00Z'
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number, starting at 1.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
      description: Number of records per page.
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
            message: Missing or invalid API key
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: QUERY_FAILED
            message: Database query failed
  schemas:
...
    DbTestResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        message:
          type: string
          example: Database connection successful
...
    PagedResponseBase:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 100
        count:
          type: integer
          example: 1
    CompanyList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Company'
    Company:
      type: object
      properties:
        companyId:
          type: string
          example: '001'
        companyName:
          type: string
          example: Summit Interconnect
    GLAccountList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/GLAccount'
...
    CustomerList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
...
    VendorList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Vendor'
...
    PurchaseOrderList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PurchaseOrder'
...
    ReceiptList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Receipt'
...
    InvoiceList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Invoice'
...
    StatementList:
      allOf:
        - $ref: '#/components/schemas/PagedResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Statement'
...
    QuadientInvoiceRequest:
      type: object
      required:
        - invoiceNumber
        - invoiceDate
        - dueDate
        - totalAmount
        - lines
      properties:
        invoiceNumber:
          type: string
          maxLength: 50
          example: INV-12345
        vendorKey:
          type: integer
          nullable: true
          description: Sage 500 tapVendor.VendKey. Preferred vendor identifier for Sage/DIM processing.
          example: 123
        vendorId:
          type: string
          nullable: true
          maxLength: 50
          description: Human-readable Sage vendor ID. Useful for audit/debugging.
          example: INSUL
        invoiceDate:
          type: string
          format: date
          example: '2026-05-07'
        dueDate:
          type: string
          format: date
          example: '2026-06-06'
        memo:
          type: string
          nullable: true
          example: 3-way matched invoice from Beanworks
        beanworksInvoiceUrl:
          type: string
          nullable: true
          format: uri
          example: https://app.beanworks.com/invoices/12345
        currency:
          type: string
          nullable: true
          maxLength: 10
          example: USD
        companyId:
          type: string
          nullable: true
          maxLength: 20
          example: '001'
        totalAmount:
          type: number
          format: double
          example: 125.00
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuadientInvoiceLineRequest'
      description: One invoice per POST request. Either vendorKey or vendorId must be supplied; vendorKey is preferred for Sage/DIM processing.
...
    QuadientInvoiceResponse:
      type: object
      properties:
        status:
          type: string
          example: received
        processingStatus:
          type: string
          example: ReadyForDIM
        stagingId:
          type: integer
          example: 101
        invoiceNumber:
          type: string
          example: INV-12345
        vendorKey:
          type: integer
          nullable: true
          example: 123
        vendorId:
          type: string
          nullable: true
          example: INSUL
        lineCount:
          type: integer
          example: 3
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
          example: Invoice payload failed validation
        details:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: QUERY_FAILED
        message:
          type: string
          example: Error details